Sematic_based/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
google-genai
requests
pydantic

